refactor: update import paths for monorepo structure
