+++ packages/mindweaver/internal/mind/bootstrap/bootstrap.go
@@ -8,16 +8,16 @@ import (
 	"fmt"
 	"log/slog"
 
-	"github.com/nkapatos/mindweaver/internal/mind/collections"
-	"github.com/nkapatos/mindweaver/internal/mind/links"
-	"github.com/nkapatos/mindweaver/internal/mind/meta"
-	"github.com/nkapatos/mindweaver/internal/mind/notes"
-	"github.com/nkapatos/mindweaver/internal/mind/notetypes"
-	"github.com/nkapatos/mindweaver/internal/mind/search"
-	"github.com/nkapatos/mindweaver/internal/mind/store"
-	"github.com/nkapatos/mindweaver/internal/mind/tags"
-	"github.com/nkapatos/mindweaver/internal/mind/templates"
-	mindmigrations "github.com/nkapatos/mindweaver/migrations/mind"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/collections"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/links"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/meta"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/notes"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/notetypes"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/search"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/tags"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/templates"
+	mindmigrations "github.com/nkapatos/mindweaver/packages/mindweaver/migrations/mind"
 
 	"github.com/labstack/echo/v4"
 	_ "modernc.org/sqlite"
