+++ packages/mindweaver/internal/mind/collections/collections_services.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"log/slog"
 
-	"github.com/nkapatos/mindweaver/internal/mind/store"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
 	"github.com/nkapatos/mindweaver/pkg/middleware"
 	"github.com/nkapatos/mindweaver/pkg/sqlcext"
 	"github.com/nkapatos/mindweaver/pkg/utils"
