+++ packages/mindweaver/internal/mind/links/link_services_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 
-	"github.com/nkapatos/mindweaver/internal/mind/store"
-	mindmigrations "github.com/nkapatos/mindweaver/migrations/mind"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
+	mindmigrations "github.com/nkapatos/mindweaver/packages/mindweaver/migrations/mind"
 	"github.com/nkapatos/mindweaver/pkg/testdb"
 	"github.com/nkapatos/mindweaver/pkg/utils"
 )
