+++ packages/mindweaver/internal/mind/links/links_converters_v3.go
@@ -3,8 +3,8 @@ package links
 import (
 	"fmt"
 
-	mindv3 "github.com/nkapatos/mindweaver/internal/mind/gen/v3"
-	"github.com/nkapatos/mindweaver/internal/mind/store"
+	mindv3 "github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
