+++ packages/mindweaver/internal/mind/links/links_handlers_v3.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"connectrpc.com/connect"
-	mindv3 "github.com/nkapatos/mindweaver/internal/mind/gen/v3"
-	"github.com/nkapatos/mindweaver/internal/mind/gen/v3/mindv3connect"
+	mindv3 "github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3"
+	"github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3/mindv3connect"
 	"github.com/nkapatos/mindweaver/pkg/pagination"
 )
