+++ packages/mindweaver/internal/mind/meta/handlers_v3.go
@@ -5,8 +5,8 @@ import (
 	"context"
 
 	"connectrpc.com/connect"
-	mindv3 "github.com/nkapatos/mindweaver/internal/mind/gen/v3"
-	"github.com/nkapatos/mindweaver/internal/mind/gen/v3/mindv3connect"
+	mindv3 "github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3"
+	"github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3/mindv3connect"
 )
 
 type NoteMetaHandlerV3 struct {
