+++ packages/mindweaver/internal/mind/notes/note_services.go
@@ -6,11 +6,11 @@ import (
 	"fmt"
 	"log/slog"
 
-	"github.com/nkapatos/mindweaver/internal/mind/links"
-	"github.com/nkapatos/mindweaver/internal/mind/meta"
-	"github.com/nkapatos/mindweaver/internal/mind/scheduler"
-	"github.com/nkapatos/mindweaver/internal/mind/store"
-	"github.com/nkapatos/mindweaver/internal/mind/tags"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/links"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/meta"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/scheduler"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/tags"
 	"github.com/nkapatos/mindweaver/pkg/dberrors"
 	"github.com/nkapatos/mindweaver/pkg/markdown"
 	"github.com/nkapatos/mindweaver/pkg/middleware"
