+++ packages/mindweaver/internal/mind/notes/notes_handlers_v3.go
@@ -7,12 +7,12 @@ import (
 	"strconv"
 
 	"connectrpc.com/connect"
-	mindv3 "github.com/nkapatos/mindweaver/internal/mind/gen/v3"
-	"github.com/nkapatos/mindweaver/internal/mind/gen/v3/mindv3connect"
-	"github.com/nkapatos/mindweaver/internal/mind/links"
-	"github.com/nkapatos/mindweaver/internal/mind/meta"
-	"github.com/nkapatos/mindweaver/internal/mind/store"
-	"github.com/nkapatos/mindweaver/internal/mind/tags"
+	mindv3 "github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3"
+	"github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3/mindv3connect"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/links"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/meta"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/tags"
 	"github.com/nkapatos/mindweaver/pkg/dberrors"
 	"github.com/nkapatos/mindweaver/pkg/pagination"
 	"github.com/nkapatos/mindweaver/pkg/utils"
