+++ packages/mindweaver/internal/mind/tags/tags_handlers_v3.go
@@ -4,9 +4,9 @@ import (
 	"context"
 
 	"connectrpc.com/connect"
-	mindv3 "github.com/nkapatos/mindweaver/internal/mind/gen/v3"
-	"github.com/nkapatos/mindweaver/internal/mind/gen/v3/mindv3connect"
-	"github.com/nkapatos/mindweaver/internal/mind/store"
+	mindv3 "github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3"
+	"github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3/mindv3connect"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
 	"github.com/nkapatos/mindweaver/pkg/pagination"
 )
