+++ packages/mindweaver/internal/mind/tags/tags_services.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"log/slog"
 
-	"github.com/nkapatos/mindweaver/internal/mind/store"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
 	"github.com/nkapatos/mindweaver/pkg/dberrors"
 	"github.com/nkapatos/mindweaver/pkg/middleware"
 )
