+++ packages/mindweaver/internal/mind/templates/templates_converters_v3.go
@@ -3,8 +3,8 @@ package templates
 import (
 	"fmt"
 
-	mindv3 "github.com/nkapatos/mindweaver/internal/mind/gen/v3"
-	"github.com/nkapatos/mindweaver/internal/mind/store"
+	mindv3 "github.com/nkapatos/mindweaver/pkg/gen/proto/mind/v3"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
 	"github.com/nkapatos/mindweaver/pkg/utils"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
