+++ packages/mindweaver/internal/mind/testing/testdb.go
@@ -6,8 +6,8 @@ import (
 	"log/slog"
 	"testing"
 
-	"github.com/nkapatos/mindweaver/internal/mind/store"
-	"github.com/nkapatos/mindweaver/migrations/mind"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/internal/mind/store"
+	"github.com/nkapatos/mindweaver/packages/mindweaver/migrations/mind"
 	"github.com/nkapatos/mindweaver/pkg/testdb"
 )
